+++ contribs/dvrp/src/main/java/org/matsim/contrib/dvrp/analysis/VehicleOccupancyProfileView.java
@@ -70,7 +70,7 @@ public class VehicleOccupancyProfileView implements ProfileWriter.ProfileView {
 	}
 
 	@Override
-	public int[] times() {
+	public double[] times() {
 		return calculator.getTimeDiscretizer().getTimes();
 	}
 }
